pyproject.toml
[tool.poetry]
name = "vmpro1"
version = "0.1.0"
description = "Modern Data Engineering Platform - Optimized for Low Resource Environments"
authors = ["venysssssssssss <viniciussport2004@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
apache-airflow = "2.8.0"
apache-airflow-providers-postgres = "^5.10.0"
pandas = "^2.1.4"
faker = "^22.0.0"
sqlalchemy = "^1.4.48"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"
great-expectations = "^0.18.8"
pyspark = "^3.5.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
